ER Diagram
 The Er diagram below represents a college management system
database. The ER diagram shows all of the tables and the relationships between them. The
entities are student, degree, department, professor, course, classroom, building, assistant
professor, and advisor. We have the relations between them shown above as well. Each entity
previously described has their own primary key and foreign key. In this case, the bold text is the
primary key and the foreign key is italic. There is also one to one relationships, one to many, and
many to many. Student’s for this diagram only get one degree so the relationship is one to one.
Student has one degree, and a degree can have many students. For a one-to-many relationship.
One Professor can teach many courses. For many-to-many relationships. Many students have
many professors.